pyproject.toml
[project]
name = "caihub-backend"
version = "0.1.0"
description = "Initial FastAPI backend architecture for the CaiHub platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0,<2.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "numpy>=2.1.0,<3.0.0",
    "opencv-python-headless>=4.10.0,<5.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "sqlalchemy>=2.0.36,<3.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
